Code upgrade
In this massive update i have changed most of the elements in this project. Using textscript i fully changed the ui and the code, making the ui more complex and beautiful at the same time. Using tsx as my new base has opened a lot of opportunities for me to develop this project further